requirements.txt
boto3==1.18.58
botocore==1.21.58
certifi==2021.5.30
click==8.0.2
jmespath==0.10.0
python-dateutil==2.8.2
PyYAML==5.4.1
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.7
